+++ source/codegen/metadata/nifake/CHANGES.md
@@ -7,7 +7,9 @@ The following functions, not originally in nimi-python metadata, were newly adde
 - `'CloseExtCal'`
 - `'BoolArrayInputFunction'`
 	- This function allows testing of ViBoolean[] input parameter
- 
+- `'GetPatternPinIndexes'` 
+  -  This function allows testing of ivi-dance-with-twist mechanism
+
 The following functions were tagged with `'init_method': True,` to ensure their generated service handlers register the new session
 with the session_repository.
 - `InitWithOptions`